reorganize test_utils
- move several test helpers from test_processing.py to processing_helpers.py
- update imports to reflect these changes
